# docs/platform/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TRUST platform-api",
    "version": "m3a1"
  },
  "paths": {
...
    "/api/v1/assets": {
      "post": {
        "summary": "Upload de un Asset (multipart: campos primero, `file` último); límite por actor (BL-10)",
        "x-trust-roles": [
          "OPERATOR",
          "ADMIN"
        ],
        "responses": {
          "200": {
            "description": "Asset",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Asset"
                }
              }
            }
          },
          "201": {
            "description": "Asset",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Asset"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 200
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "session": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "surfaceType": {
                    "type": "string"
                  },
                  "requiredDurationMs": {
                    "type": "string"
                  },
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "surfaceType",
                  "file"
                ]
              },
              "encoding": {
                "file": {
                  "contentType": "application/octet-stream"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Assets del actor (cursor)",
        "x-trust-roles": [
          "OPERATOR",
          "ADMIN"
        ],
        "responses": {
          "200": {
            "description": "AssetList",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetList"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "UPLOADING",
                "VALIDATING",
                "READY",
                "REJECTED"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/v1/assets/{id}": {
      "get": {
        "summary": "Un Asset del actor",
        "x-trust-roles": [
          "OPERATOR",
          "ADMIN"
        ],
        "responses": {
          "200": {
            "description": "Asset",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Asset"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/v1/assets/{id}/status": {
      "get": {
        "summary": "Estado de un Asset",
        "x-trust-roles": [
          "OPERATOR",
          "ADMIN"
        ],
        "responses": {
          "200": {
            "description": "AssetStatus",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetStatus"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "minLength": 1
          },
          "message": {
            "type": "string",
            "minLength": 1
          },
          "details": {
            "type": "object",
            "additionalProperties": {}
          },
          "requestId": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "code",
          "message",
          "requestId"
        ]
      },
...
      "Asset": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "UPLOADING",
              "VALIDATING",
              "READY",
              "REJECTED"
            ]
          },
          "surfaceType": {
            "type": "string"
          },
          "originalFilename": {
            "type": "string"
          },
          "sha256": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[0-9a-f]{64}$"
              },
              {
                "type": "null"
              }
            ]
          },
          "sizeBytes": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "mimeType": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "container": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "width": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "height": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "fps": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "codec": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "durationMs": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "storedObjectId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ]
          },
          "rejection": {
            "anyOf": [
              {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string",
                    "enum": [
                      "ASSET_TOO_LARGE",
                      "ASSET_EMPTY",
                      "ASSET_BAD_CONTAINER",
                      "ASSET_BAD_CODEC",
                      "ASSET_BAD_RESOLUTION",
                      "ASSET_BAD_FPS",
                      "ASSET_TOO_SHORT",
                      "ASSET_NO_VIDEO",
                      "ASSET_MULTIPLE_VIDEO_STREAMS",
                      "ASSET_CORRUPT",
                      "ASSET_INSPECTION_TIMEOUT",
                      "ASSET_STORAGE_ERROR"
                    ]
                  },
                  "message": {
                    "type": "string",
                    "minLength": 1
                  },
                  "details": {
                    "type": "object",
                    "additionalProperties": {}
                  },
                  "remediation": {
                    "type": "object",
                    "properties": {
                      "summary": {
                        "type": "string",
                        "minLength": 1
                      },
                      "ffmpeg": {
                        "type": "object",
                        "properties": {
                          "args": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "displayCommand": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "args",
                          "displayCommand"
                        ]
                      }
                    },
                    "required": [
                      "summary"
                    ]
                  }
                },
                "required": [
                  "code",
                  "message",
                  "remediation"
                ]
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "status",
          "surfaceType",
          "originalFilename",
          "sha256",
          "sizeBytes",
          "mimeType",
          "container",
          "width",
          "height",
          "fps",
          "codec",
          "durationMs",
          "storedObjectId",
          "rejection"
        ]
      },
      "AssetStatus": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "UPLOADING",
              "VALIDATING",
              "READY",
              "REJECTED"
            ]
          },
          "rejection": {
            "anyOf": [
              {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string",
                    "enum": [
                      "ASSET_TOO_LARGE",
                      "ASSET_EMPTY",
                      "ASSET_BAD_CONTAINER",
                      "ASSET_BAD_CODEC",
                      "ASSET_BAD_RESOLUTION",
                      "ASSET_BAD_FPS",
                      "ASSET_TOO_SHORT",
                      "ASSET_NO_VIDEO",
                      "ASSET_MULTIPLE_VIDEO_STREAMS",
                      "ASSET_CORRUPT",
                      "ASSET_INSPECTION_TIMEOUT",
                      "ASSET_STORAGE_ERROR"
                    ]
                  },
                  "message": {
                    "type": "string",
                    "minLength": 1
                  },
                  "details": {
                    "type": "object",
                    "additionalProperties": {}
                  },
                  "remediation": {
                    "type": "object",
                    "properties": {
                      "summary": {
                        "type": "string",
                        "minLength": 1
                      },
                      "ffmpeg": {
                        "type": "object",
                        "properties": {
                          "args": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "displayCommand": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "args",
                          "displayCommand"
                        ]
                      }
                    },
                    "required": [
                      "summary"
                    ]
                  }
                },
                "required": [
                  "code",
                  "message",
                  "remediation"
                ]
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "status",
          "rejection"
        ]
      },
...
    },
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-trust_session",
        "description": "Cookie de sesión server-side (C1). En desarrollo por http el nombre es trust_session."
      }
    }
  }
}